fix: remove bpf dependency in e2e tests
