+++ internal/eventhandler/utils/utils.go
@@ -1,0 +1,37 @@
+package utils
+
+import (
+	"fmt"
+
+	"k8s.io/apimachinery/pkg/util/validation"
+)
+
+// GetWorkloadPolicyProposalName returns the name of WorkloadPolicyProposal
+// based on a high level resource and its name.
+func GetWorkloadPolicyProposalName(kind string, resourceName string) (string, error) {
+	var shortname string
+	switch kind {
+	case "Deployment":
+		shortname = "deploy"
+	case "ReplicaSet":
+		shortname = "rs"
+	case "DaemonSet":
+		shortname = "ds"
+	case "CronJob":
+		shortname = "cronjob"
+	case "Job":
+		shortname = "job"
+	case "StatefulSet":
+		shortname = "sts"
+	default:
+		return "", fmt.Errorf("unknown kind: %s", kind)
+	}
+	ret := shortname + "-" + resourceName
+
+	// The max name length in k8s
+	if len(ret) > validation.DNS1123SubdomainMaxLength {
+		return "", fmt.Errorf("the name %s exceeds the maximum name length", ret)
+	}
+
+	return shortname + "-" + resourceName, nil
+}
